Rename the Altaibot demo
Got feedback that the name was confusing because it looks like Alt + AI
