+++ aiohttp_poe/src/aiohttp_poe/__main__.py
@@ -1,5 +1,5 @@
 from aiohttp_poe.base import run
-from aiohttp_poe.samples.altai import AltaiHandler
+from aiohttp_poe.samples.catbot import CatBotHandler
 
 if __name__ == "__main__":
-    run(AltaiHandler())
+    run(CatBotHandler())
