+++ fastapi_poe/README.md
@@ -28,4 +28,4 @@ if __name__ == "__main__":
 ```
 
 For a more advanced example that exercises more of the Poe protocol, see
-[Altaibot](./src/fastapi_poe/samples/altai.py).
+[Catbot](./src/fastapi_poe/samples/catbot.py).
