+++ fastapi_poe/src/fastapi_poe/__main__.py
@@ -1,5 +1,5 @@
 from fastapi_poe import run
-from fastapi_poe.samples.altai import AltaiHandler
+from fastapi_poe.samples.catbot import CatBotHandler
 
 if __name__ == "__main__":
-    run(AltaiHandler())
+    run(CatBotHandler())
